refactored Program
